+++ src/main/resources/static/js/profile.js
@@ -1,8 +1,69 @@
-// 쿠키값 가져오기
-let token = Cookies.get('Authorization');
 
-const updateForm = document.getElementById("form2");
-updateForm.addEventListener("submit", (e) => e.preventDefault());
+// 비밀번호 내용 표시 및 숨기기
+const removePasswordInput = document.getElementById('removePassword');
+const oldPasswordInput = document.getElementById('oldPassword');
+const newPasswordInput = document.getElementById('newPassword');
+const showRemovePwBtn = document.getElementById('showRemovePwBtn');
+const showOldPwBtn = document.getElementById('showOldPwBtn');
+const showNewPwBtn = document.getElementById('showNewPwBtn');
+
+showRemovePwBtn.addEventListener('click', (event) => {
+  event.preventDefault();
+
+  if (removePasswordInput.type === 'password') {
+    removePasswordInput.type = 'text';
+    //removePasswordInput.classList.add('visible'); // Apply the 'visible' class
+  } else {
+    removePasswordInput.type = 'password';
+    //removePasswordInput.classList.remove('visible'); // Remove the 'visible' class
+  }
+});
+
+showOldPwBtn.addEventListener('click', (event) => {
+  event.preventDefault();
+
+  if (oldPasswordInput.type === 'password') {
+    oldPasswordInput.type = 'text';
+    //removePasswordInput.classList.add('visible'); // Apply the 'visible' class
+  } else {
+    oldPasswordInput.type = 'password';
+    //removePasswordInput.classList.remove('visible'); // Remove the 'visible' class
+  }
+});
+
+showNewPwBtn.addEventListener('click', (event) => {
+  event.preventDefault();
+
+  if (newPasswordInput.type === 'password') {
+    newPasswordInput.type = 'text';
+    //removePasswordInput.classList.add('visible'); // Apply the 'visible' class
+  } else {
+    newPasswordInput.type = 'password';
+    //removePasswordInput.classList.remove('visible'); // Remove the 'visible' class
+  }
+});
+
+
+
+const updateBtn = document.getElementById("update");
+const removeBtn = document.getElementById("remove");
+const fistForm = document.getElementById("form1");
+const secondForm = document.getElementById("form2");
+const container = document.querySelector(".container");
+
+updateBtn.addEventListener("click", () => {
+  container.classList.remove("right-panel-active");
+});
+
+removeBtn.addEventListener("click", () => {
+  container.classList.add("right-panel-active");
+});
+
+fistForm.addEventListener("submit", (e) => e.preventDefault());
+secondForm.addEventListener("submit", (e) => e.preventDefault());
+
+
+
 
 // alert 창 표시를 위한 변수 선언
 const Toast = Swal.mixin({
@@ -16,6 +77,9 @@ const Toast = Swal.mixin({
     toast.addEventListener('mouseleave', Swal.resumeTimer)
   }
 })
+
+// 쿠키값 가져오기
+let token = Cookies.get('Authorization');
 
 $(document).ready(function () {
 
@@ -58,7 +122,8 @@ function getUser(token) {
   if (username === "") { // 페이로드 추출이 안될 경우
     removeToken();
   } else {
-    document.getElementById("username").value = username;
+    document.getElementById("updateUsername").value = username;
+    document.getElementById("removeUsername").value = username;
   }
 }
 
@@ -84,7 +149,7 @@ function removeToken() {
   Cookies.remove('Authorization', {path: '/'});
   Swal.fire({
     icon: 'warning',
-    title: '로그인 만료',
+    title: '로그인 요망',
     text: '인증이 만료되어 재로그인 부탁드립니다.',
   }).then(function () {
     window.location.href = "/main";
@@ -93,8 +158,8 @@ function removeToken() {
 
 function updateUser() {
   let email = document.getElementById('email').value;
-  let oldPassword = document.getElementById('old-password').value;
-  let newPassword = document.getElementById('new-password').value;
+  let oldPassword = document.getElementById('oldPassword').value;
+  let newPassword = document.getElementById('newPassword').value;
 
   const RegExp2 = /^[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/;
   const num = newPassword.search(/[0-9]/g);
@@ -180,18 +245,33 @@ function updateUser() {
 }
 
 function removeUser() {
+  let removePassword = document.getElementById('removePassword').value;
+
+  if (removePassword.search(/\s/) != -1) {
+    Swal.fire({
+      icon: 'warning',
+      title: '비밀번호 입력 오류',
+      text: '비밀번호는 공백 없이 입력해주세요.',
+    });
+    $('#removePassword').focus();
+    return false;
+  }
 
   $.ajax({
     type: "DELETE",
     url: `/api/users`,
     contentType: "application/json",
+    data: JSON.stringify({
+      password: removePassword
+    }),
   })
   .done(function (res, status, xhr) {
     Toast.fire({
       icon: 'success',
       title: '회원을 탈퇴하였습니다.'
     }).then(function () {
-      removeToken();
+      Cookies.remove('Authorization', {path: '/'});
+      window.location.href="/main";
     })
   })
   .fail(function (jqXHR, textStatus, error) {
